Test Specific Variables
The user can now populate test specific variables! This was a major obstacle for the past couple of days. For some reason, we couldn't pass both the form as an object and the form id as a string into populateTestSpecificInformation(). After fiddling with JSON and not having any luck, I discovered that I could add an invisible input to the HTML form for each test, enabling us to determine what form is passed into populateTestSpecificInformation().